+++ src/ethereum/contracts.js
@@ -22,9 +22,9 @@ function Contracts() {
    * @param {string} address
    * @returns {string} - The index of the address saved
    */
-  this.addresses.set = (name, address) => {
+  this.addresses.set = (name, address, links) => {
     addresses.path = this.paths.address;
-    return addresses.set(name, address);
+    return addresses.set(name, address, links);
   };
 
   /**
@@ -39,14 +39,36 @@ function Contracts() {
   };
 
   /**
+   * Get the links of a deployed contract
+   * @param {string} name
+   * @param {number} index
+   * @returns {Object}
+   */
+  this.addresses.getLinks = (name, index) => {
+    addresses.path = this.paths.address;
+    return addresses.getLinks(name, index);
+  };
+
+  /**
    * Get all the deployed addresses of a contract
    * @param {string} name
-   * @returns {Array}
+   * @returns {Array} - Arr
    */
   this.addresses.getAll = (name) => {
     addresses.path = this.paths.address;
     return addresses.getAll(name);
   };
+
+  /**
+   * Get all the deployed address links of a contract
+   * @param {string} name
+   * @returns {Array} - Array of objects containing the contract address and link addresses
+   */
+  this.addresses.getAllLinks = (name) => {
+    addresses.path = this.paths.address;
+    return addresses.getAllLinks(name);
+  };
+
 }
 
 const contracts = new Contracts();