+++ testing/tests/core.js
@@ -25,21 +25,21 @@ const xtest = (describe, callback) => {
 };
 
 test('Initializing connection', t => {
-  delib.eth.init();
+  delib.init();
 
-  t.equal(delib.eth.checkConnection('rpc'), true, 'Expect checkConnection rpc to be true');
+  t.equal(delib.checkConnection('rpc'), true, 'Expect checkConnection rpc to be true');
 
-  delib.eth.initIPC('./chain1/devchain/geth.ipc');
+  delib.initIPC('./chain1/devchain/geth.ipc');
 
-  t.equal(delib.eth.checkConnection('ipc'), true, 'Expect checkConnect ipc to be true');
+  t.equal(delib.checkConnection('ipc'), true, 'Expect checkConnect ipc to be true');
 
-  delib.eth.closeIPC();
+  delib.closeIPC();
 
   t.end();
 });
 
 test('Building Bank contract', t => {
-  delib.eth.build(['Bank', 'BadBank'])
+  delib.build(['Bank', 'BadBank'])
     .then(contracts => {
       t.equal(contracts[0], 'BadBank', 'Expect first contract to be BadBank');
       t.equal(contracts[1], 'Bank', 'Expect second contract to be Bank');
@@ -52,14 +52,14 @@ test('Building Bank contract', t => {
 });
 
 test('Deploying Bank contract with gas estimate', t => {
-  delib.eth.init();
+  delib.init();
 
-  delib.eth.deploy.estimate('Bank')
+  delib.deploy.estimate('Bank')
     .then(estimate => {
       t.equal(estimate > 100000, true, 'Expect deploy.estimate to return an estimate greater than 100000');
       t.equal(estimate < 500000, true, 'Expect deploy.estimate to return an estimate less than 500000');
 
-      return delib.eth.deploy('Bank');
+      return delib.deploy('Bank');
     })
     .then(instance => {
       t.notEqual(instance.address, undefined, 'Expect deploy to return an instance with an address property');
@@ -72,9 +72,9 @@ test('Deploying Bank contract with gas estimate', t => {
 });
 
 test('Deploying Bank contract with no gas estimate', t => {
-  delib.eth.init();
+  delib.init();
 
-  delib.eth.deploy('Bank', [], {gas: 3000000})
+  delib.deploy('Bank', [], {gas: 3000000})
     .then(instance => {
       t.notEqual(instance.address, undefined, 'Expect deploy to return an instance with an address property');
       t.end();
@@ -86,10 +86,10 @@ test('Deploying Bank contract with no gas estimate', t => {
 });
 
 test('Executing Bank contract methods with gas estimate', t => {
-  delib.eth.init();
-  delib.eth.deploy('Bank')
+  delib.init();
+  delib.deploy('Bank')
     .then(instance => {
-      return delib.eth.exec('Bank').estimate.deposit({
+      return delib.exec('Bank').estimate.deposit({
         value: 4
       });
     })
@@ -97,21 +97,21 @@ test('Executing Bank contract methods with gas estimate', t => {
       t.equal(estimate > 30000, true, 'Expect exec.estimate to return a gas estimate greater than 30000');
       t.equal(estimate < 200000, true, 'Expect exec.estimate to return a gas estimate less than 200000');
 
-      return delib.eth.exec('Bank').deposit({
+      return delib.exec('Bank').deposit({
         value: 4
       });
     })
     .then(tx => {
       t.ok(tx, 'Expect deposit method with a value of 4 to return a tx response');
-      return delib.eth.exec('Bank').call.checkAmount();
+      return delib.exec('Bank').call.checkAmount();
     })
     .then(amount => {
       t.equal(Number(amount), 4, 'Expect exec.call checkAmount method to return 4');
-      return delib.eth.exec('Bank').withdraw(0);
+      return delib.exec('Bank').withdraw(0);
     })
     .then(tx => {
       t.ok(tx, 'Expect withdraw method to return a tx response');
-      return delib.eth.exec('Bank').checkAmount();
+      return delib.exec('Bank').checkAmount();
     })
     .then(amount => {
       t.equal(Number(amount), 0, 'Expect checkAmount method to return 0 now');
@@ -124,18 +124,18 @@ test('Executing Bank contract methods with gas estimate', t => {
 });
 
 test('Executing Bank contract methods with no gas estimate', t => {
-  delib.eth.init();
+  delib.init();
   const gas = 500000;
-  delib.eth.deploy('Bank')
+  delib.deploy('Bank')
     .then(instance => {
-      return delib.eth.exec('Bank').deposit({
+      return delib.exec('Bank').deposit({
         value: 4,
         gas: gas
       });
     })
     .then(tx => {
       t.ok(tx, 'Expect deposit method with a value of 4 to return a tx response');
-      return delib.eth.exec('Bank').checkAmount({gas: gas});
+      return delib.exec('Bank').checkAmount({gas: gas});
     })
     .then(amount => {
       t.equal(Number(amount), 4, 'Expect checkAmount method to return 4');
@@ -148,14 +148,14 @@ test('Executing Bank contract methods with no gas estimate', t => {
 });
 
 test('Getting Bank contract event logs', t => {
-  delib.eth.init();
+  delib.init();
 
-  delib.eth.deploy('Bank')
+  delib.deploy('Bank')
     .then(instance => {
       const promises = [
-        delib.eth.exec('Bank').deposit({value: 3, account: 0}),
-        delib.eth.exec('Bank').deposit({value: 4, account: 0}),
-        delib.eth.exec('Bank').deposit({value: 5, account: 0}),
+        delib.exec('Bank').deposit({value: 3, account: 0}),
+        delib.exec('Bank').deposit({value: 4, account: 0}),
+        delib.exec('Bank').deposit({value: 5, account: 0}),
       ];
 
       console.log('    - Executing deposit 9 times');
@@ -163,27 +163,27 @@ test('Getting Bank contract event logs', t => {
     })
     .then(txs => {
       const promises = [
-        delib.eth.exec('Bank').deposit({value: 6, account: 0}),
-        delib.eth.exec('Bank').deposit({value: 3, account: 1}),
-        delib.eth.exec('Bank').deposit({value: 4, account: 1}),
+        delib.exec('Bank').deposit({value: 6, account: 0}),
+        delib.exec('Bank').deposit({value: 3, account: 1}),
+        delib.exec('Bank').deposit({value: 4, account: 1}),
       ];
 
       return Promise.all(promises);
     })
     .then(txs => {
       const promises = [
-        delib.eth.exec('Bank').deposit({value: 5, account: 1}),
-        delib.eth.exec('Bank').deposit({value: 3, account: 2}),
-        delib.eth.exec('Bank').deposit({value: 4, account: 2}),
+        delib.exec('Bank').deposit({value: 5, account: 1}),
+        delib.exec('Bank').deposit({value: 3, account: 2}),
+        delib.exec('Bank').deposit({value: 4, account: 2}),
       ];
       return Promise.all(promises);
     })
     .then(txs => {
 
       const promises = [
-        delib.eth.exec('Bank').withdraw(0, {account: 0}),
-        delib.eth.exec('Bank').withdraw(0, {account: 1}),
-        delib.eth.exec('Bank').withdraw(0, {account: 2}),
+        delib.exec('Bank').withdraw(0, {account: 0}),
+        delib.exec('Bank').withdraw(0, {account: 1}),
+        delib.exec('Bank').withdraw(0, {account: 2}),
       ];
 
       console.log('    - Executing withdraw 3 times');
@@ -191,17 +191,17 @@ test('Getting Bank contract event logs', t => {
       return Promise.all(promises);
     })
     .then(logs => {
-      return delib.eth.events('Bank', 'allEvents', 'all');
+      return delib.events('Bank', 'allEvents', 'all');
     })
     .then(logs => {
       t.equal(logs.length, 12, 'Expect allEvents to return 12 logs');
 
-      return delib.eth.events('Bank', 'depositEvent', 'all');
+      return delib.events('Bank', 'depositEvent', 'all');
     })
     .then(logs => {
       t.equal(logs.length, 9, 'Expect the depositEvent to return 9 logs');
 
-      return delib.eth.events('Bank', 'withdrawEvent', 'all');
+      return delib.events('Bank', 'withdrawEvent', 'all');
     })
     .then(logs => {
       t.equal(logs.length, 3, 'Expect withdrawEvent to return 3 logs');
@@ -210,7 +210,7 @@ test('Getting Bank contract event logs', t => {
         event: 'withdrawEvent'
       };
 
-      return delib.eth.events('Bank', 'allEvents', 'all', filter);
+      return delib.events('Bank', 'allEvents', 'all', filter);
     })
     .then(logs => {
       t.equal(logs.length, 3, 'Expect allEvents to return 9 logs with the filter object { event: withdrawEvent }');
@@ -228,10 +228,10 @@ test('Getting Bank contract event logs', t => {
         }
       };
 
-      return delib.eth.events('Bank', 'depositEvent', 'all', filter);
+      return delib.events('Bank', 'depositEvent', 'all', filter);
     })
     .then(logs => {
-      const web3 = delib.eth.init();
+      const web3 = delib.init();
 
       t.equal(logs.length, 2, 'Expect depositEvent to return 2 logs with the filter object { args: { _amount: callback for 5 } }');
 
@@ -248,7 +248,7 @@ test('Getting Bank contract event logs', t => {
         }
       };
 
-      return delib.eth.events('Bank', 'depositEvent', 'all', filter);
+      return delib.events('Bank', 'depositEvent', 'all', filter);
     })
     .then(logs => {
       t.equal(logs.length, 2, 'Expect depositEvent to return 2 logs with the filter object { args: { _user: [acc0, acc1], _amount: callback for 3 } }');
@@ -264,10 +264,10 @@ test('Getting Bank contract event logs', t => {
 test('Watching for Bank contract event logs', t => {
   const depositLogs = [];
   const withdrawLogs = [];
-  const web3 = delib.eth.init();
+  const web3 = delib.init();
   let withdrawWatch;
   let depositWatch;
-  delib.eth.deploy('Bank')
+  delib.deploy('Bank')
     .then(instance => {
 
       const filter = {
@@ -282,7 +282,7 @@ test('Watching for Bank contract event logs', t => {
         }
       };
 
-      depositWatch = delib.eth.watch('Bank', 'depositEvent', filter, (err, log) => {
+      depositWatch = delib.watch('Bank', 'depositEvent', filter, (err, log) => {
         if (err) {
           console.error(err);
           t.fail();
@@ -292,7 +292,7 @@ test('Watching for Bank contract event logs', t => {
 
       });
 
-      withdrawWatch = delib.eth.watch('Bank', 'withdrawEvent', filter, (err, log) => {
+      withdrawWatch = delib.watch('Bank', 'withdrawEvent', filter, (err, log) => {
         if (err) {
           console.error(err);
           t.fail();
@@ -303,25 +303,25 @@ test('Watching for Bank contract event logs', t => {
 
       console.log('    - Performing deposits and withdraws');
 
-      return delib.eth.exec('Bank').deposit({value: 10});
+      return delib.exec('Bank').deposit({value: 10});
     })
     .then((tx) => {
-      return delib.eth.exec('Bank').withdraw(10);
+      return delib.exec('Bank').withdraw(10);
     })
     .then((tx) => {
-      return delib.eth.exec('Bank').deposit({value: 3});
+      return delib.exec('Bank').deposit({value: 3});
     })
     .then((tx) => {
-      return delib.eth.exec('Bank').deposit({value: 4});
+      return delib.exec('Bank').deposit({value: 4});
     })
     .then((tx) => {
-      return delib.eth.exec('Bank').withdraw(4);
+      return delib.exec('Bank').withdraw(4);
     })
     .then((tx) => {
-      return delib.eth.exec('Bank').withdraw(3);
+      return delib.exec('Bank').withdraw(3);
     })
     .then(tx => {
-      return delib.eth.exec('Bank').withdraw(1);
+      return delib.exec('Bank').withdraw(1);
     })
     .then((tx) => {
       t.equal(depositLogs.length, 2, 'Expect 2 deposit logs');
@@ -342,3 +342,61 @@ test('Watching for Bank contract event logs', t => {
       console.error(err);
     });
 });
+
+
+/** N/A Tests */
+/**
+xtest('Contract JSON addresses', t => {
+  delib.init();
+  delib.deploy('ConvertLib', [], {gas: 500000})
+    .then(instance => {
+      const convertAddresses = instance.address;
+
+      const links = {
+        ConvertLib: '0xe3ad5a4775e7cdf6bf1867f20c1cbc7d192df5bb'
+      };
+
+      delib.link('Bank2', links);
+      return delib.deploy('Bank2', [], {gas: 500000});
+    })
+    .then(instance => {
+      const address = delib.contracts.addresses.get('Bank2');
+      const links = delib.contracts.addresses.getLinks('Bank2');
+      const addressesArray = delib.contracts.addresses.getAll('Bank2');
+      const linksArray = delib.contracts.addresses.getAllLinks('Bank2');
+
+      const address2 = delib.contracts.addresses.get('Convert');
+      const links2 = delib.contracts.addresses.getLinks('Convert');
+      const addressesArray2 = delib.contracts.addresses.getAll('Convert');
+      const linksArray2 = delib.contracts.addresses.getAllLinks('Convert');
+
+      // console.log('address Bank2', address);
+      // console.log('links Bank2', links);
+      // console.log('addressesArray Bank2', addressesArray);
+      // console.log('linksArray Bank2', linksArray);
+      //
+      // console.log('address Convert', address2);
+      // console.log('links Convert', links2);
+      // console.log('addressesArray Convert', addressesArray2);
+      // console.log('linksArray Convert', linksArray2);
+
+      return delib.exec('Bank2').deposit({
+        value: 500000000000
+      });
+    })
+    .then(tx => {
+
+      return delib.exec('Bank2').call.checkAmountEther();
+    })
+    .then(amount => {
+      console.log('amount', amount);
+
+      t.end();
+      // return delib.exec('Bank2').checkAmountEther();
+    })
+    .catch(err => {
+      console.error(err);
+      t.fail();
+    });
+});
+*/